feat: done drag-n-drop

+++ src/components/burger-ingredients/burger-ingredients.jsx
@@ -3,13 +3,14 @@ import {forwardRef, useEffect, useRef, useState} from "react";
 import {Counter, CurrencyIcon, Tab} from "@ya.praktikum/react-developer-burger-ui-components";
 import PropTypes from "prop-types";
 import {useDispatch, useSelector} from "react-redux";
-import {fetchIngredients, incrementCount} from "../../services/reducers/ingredients";
+import {decrementCount, fetchIngredients} from "../../services/reducers/ingredients";
 import Preloader from "../preloader/preloader";
 import {ingredientsMapping} from "../../utils/constants";
 import {showIngredientInfo} from "../../services/reducers/ingredient-info";
 import NoContent from "../no-content/no-content";
-import {addIngredient, removeIngredient} from "../../services/reducers/burger-constructor";
+import {removeIngredient} from "../../services/reducers/burger-constructor";
 import {useDrag, useDrop} from "react-dnd";
+import {setHighlightedCategory} from "../../services/reducers/nav";
 
 const IngredientsNavigation = ({onPickCategory, currentCategory, refs}) => {
 
@@ -46,35 +47,44 @@ IngredientsNavigation.propTypes = {
 }
 
 const Ingredient = ({ingredient}) => {
-  const {image, price, name, _id, type, count} = ingredient;
+  const [isCounterActive, setIsCounterActive] = useState(true);
+  const {image, price, name, count} = ingredient;
   const dispatch = useDispatch();
-  const [{isDrag, opacity}, dragRef] = useDrag({
-    type: 'ingredient',
-    item: ingredient,
+  const [{ opacity}, dragRef] = useDrag({
+    type: 'pickedIngredient',
+    item: () => {
+      return { ingredient, ingredientType: 'new' }
+    },
     collect: monitor => ({
-      isDrag: monitor.isDragging(),
-      opacity: monitor.isDragging() ? 0.5 : 1
-    })
+      opacity: monitor.isDragging() ? 0.3 : 1,
+      isDragging: monitor.isDragging(),
+    }),
+    end() {
+      setIsCounterActive(true);
+    }
   });
 
   const handleInfoClick = () => {
     dispatch(showIngredientInfo({item: ingredient}));
   }
-  const handleAddClick = (e) => {
-    e.stopPropagation();
-    dispatch(incrementCount({_id, type}))
-    dispatch(addIngredient({item: ingredient}))
+
+  const hideCounter = () => {
+    setIsCounterActive(false);
   }
 
   return (
-    <li ref={dragRef} onClick={handleInfoClick} className={`${ingredientsStyle.ingredient} pt-6`} style={{opacity}}>
+    <li ref={dragRef}
+        onClick={handleInfoClick}
+        onMouseDown={hideCounter}
+        className={`${ingredientsStyle.ingredient} pt-6`}
+        style={{opacity}}>
       <img className={`${ingredientsStyle.ingredient__image} pl-4 pr-4`} src={image} alt={name} />
       <div className={`${ingredientsStyle.ingredient__price} pt-1 pb-1`}>
         <span className={"text text_type_digits-default"}>{price}</span>
         <CurrencyIcon type={"primary"} />
       </div>
-      <p onClick={handleAddClick} className={`${ingredientsStyle.ingredient_text} pt-2`}>{name}</p>
-      {(count && count > 0) ? <Counter count={count} size="default" extraClass={ingredientsStyle.ingredient__counter} /> : null}
+      <p className={`${ingredientsStyle.ingredient_text} pt-2`}>{name}</p>
+      {isCounterActive && (count && count > 0) ? <Counter count={count} size="default" extraClass={ingredientsStyle.ingredient__counter} /> : null}
     </li>
   )
 }
@@ -88,7 +98,7 @@ const Ingredients = forwardRef(({type}, ref) => {
   const ingredients = useSelector((state) => state.ingredients.items[type]);
 
   return (
-    <div ref={ref} className={`${ingredientsStyle.ingredients} pt-10  custom-scroll`}>
+    <div ref={ref} className={`${ingredientsStyle.ingredients} pt-10 custom-scroll`}>
       <h2>{title}</h2>
       <ul className={ingredientsStyle.ingredients__list}>
         {ingredients && ingredients.map((ingredient) => (
@@ -107,48 +117,32 @@ const BurgerIngredients = () => {
 
   const dispatch = useDispatch();
   const {status} = useSelector((state) => state.ingredients);
+  const {highlightedCategory} = useSelector((state) => state.nav);
   const ingredientContainerRef = useRef(null);
   const bunRef = useRef(null);
   const sauceRef = useRef(null);
   const mainRef = useRef(null);
-  const [highlightedCategory, setHighlightedCategory] = useState('bun');
-
+  const [currentType, setCurrentType] = useState(null);
 
   const [{isHover}, dropTarget] = useDrop({
-    accept: "ingredient",
+    accept: "pickedIngredient",
     drop(item) {
-      dispatch(removeIngredient({item: item.ingredient, index: item.index}));
+      if (item.ingredientType === 'exist') {
+        dispatch(removeIngredient({item: item.ingredient, index: item.index}));
+        dispatch(decrementCount({_id: item.ingredient._id, type: item.ingredient.type}));
+      }
+    },
+    hover(item) {
+      setCurrentType(item.ingredientType);
     },
     collect: monitor => ({
       isHover: monitor.isOver(),
     })
   });
 
-  const handleScroll = () => {
-    const elementPositions = {
-      bun: bunRef.current.offsetTop - ingredientContainerRef.current.offsetTop,
-      sauce: sauceRef.current.offsetTop - ingredientContainerRef.current.offsetTop,
-      main: mainRef.current.offsetTop - ingredientContainerRef.current.offsetTop,
-    };
-
-    const scrollTop = ingredientContainerRef.current.scrollTop;
-    let maxVisiblePosition = -1;
-    let visibleCategory = null;
-
-    for (const category in elementPositions) {
-      if (elementPositions[category] <= scrollTop && elementPositions[category] > maxVisiblePosition) {
-        maxVisiblePosition = elementPositions[category];
-        visibleCategory = category;
-      }
-    }
-
-    if (visibleCategory !== null && visibleCategory !== highlightedCategory) {
-      setHighlightedCategory(visibleCategory);
-    }
-  };
 
   const handlePickCategory = (category) => {
-    setHighlightedCategory(category);
+    dispatch(setHighlightedCategory(category));
   };
 
 
@@ -159,17 +153,46 @@ const BurgerIngredients = () => {
   }, [status, dispatch]);
 
   useEffect(() => {
+    const handleScroll = () => {
+      if (!bunRef.current || !sauceRef.current || !mainRef.current || !ingredientContainerRef.current) {
+        return;
+      }
+
+      const elementPositions = {
+        bun: bunRef.current.offsetTop - ingredientContainerRef.current.offsetTop,
+        sauce: sauceRef.current.offsetTop - ingredientContainerRef.current.offsetTop,
+        main: mainRef.current.offsetTop - ingredientContainerRef.current.offsetTop,
+      };
+
+      const scrollTop = ingredientContainerRef.current.scrollTop;
+      let maxVisiblePosition = -1;
+      let visibleCategory = null;
+
+      for (const category in elementPositions) {
+        if (elementPositions[category] <= scrollTop && elementPositions[category] > maxVisiblePosition) {
+          maxVisiblePosition = elementPositions[category];
+          visibleCategory = category;
+        }
+      }
+
+      if (visibleCategory !== null && visibleCategory !== highlightedCategory) {
+        dispatch(setHighlightedCategory(visibleCategory));
+      }
+    };
+
     const ingredientContainer = ingredientContainerRef.current;
     ingredientContainer.addEventListener('scroll', handleScroll);
     return () => ingredientContainer.removeEventListener('scroll', handleScroll);
-  });
+  }, [highlightedCategory, dispatch]);
+
+  const outline = currentType !== 'new' && isHover ? '2px solid purple' : '';
 
   return (
     <section className={`${ingredientsStyle.container}`}>
       <h1 className={"text text_type_main-large mt-10 mb-5"}>Соберите бургер</h1>
       <IngredientsNavigation currentCategory={highlightedCategory} onPickCategory={handlePickCategory}
       refs={{bun: bunRef, sauce: sauceRef, main: mainRef}}/>
-      <div ref={ingredientContainerRef} className={`${ingredientsStyle.ingredients__container} custom-scroll`}>
+      <div ref={ingredientContainerRef} className={`${ingredientsStyle.ingredients__container} custom-scroll`} style={{outline}}>
         {status === 'loading'
           ? <Preloader />
           : (status === 'failed')
